receive ALS from node

+++ lib/xlxCloudObj.cpp
@@ -104,7 +104,7 @@ BOOL CloudObjClass::UpdateHumidity(float value)
   return false;
 }
 
-BOOL CloudObjClass::UpdateBrightness(uint16_t value)
+BOOL CloudObjClass::UpdateBrightness(uint8_t value)
 {
   if( m_brightness != value ) {
     m_brightness = value;
